Add files via upload
DetectFaultyRulesinNSG.ps1: In the world which is run by technology nowadays security plays a significant role in any organization's well being. Its critical to secure our data and resources so that there is no threat. If there is a NSG in the given subscription with the rule set to Source address, Destination address, Source port and Destination port set to 'Any' then it will be a threat to the associated resources. The script helps in identifying such faulty NSGs in a given Azure subscription.